+++ src/cloudant/result.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-# Copyright (C) 2015, 2018 IBM Corp. All rights reserved.
+# Copyright (C) 2015, 2019 IBM Corp. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -454,7 +454,7 @@ class QueryResult(Result):
     :param int r: Read quorum needed for the result.  Each document is read
         from at least 'r' number of replicas before it is returned in the
         results.
-    :param str selector: Dictionary object describing criteria used to
+    :param dict selector: Dictionary object describing criteria used to
         select documents.
     :param list sort: A list of fields to sort by.  Optionally the list can
         contain elements that are single member dictionary structures that