+++ CHANGES.md
@@ -1,7 +1,9 @@
 # Unreleased
 
-- [IMPROVED] Updated `Getting started` section with a `get_query_result` example. 
+- [FIXED] Bug where document context manager performed remote save despite
+  uncaught exceptions being raised inside `with` block.
 - [FIXED] Fixed parameter type of `selector` in docstring.
+- [IMPROVED] Updated `Getting started` section with a `get_query_result` example.
 
 # 2.11.0 (2019-01-21)